Fix compilation of glew when only VS 2015 is available
Previously, glew was always compiled by VS 2013 and it worked on my machine
because it had both VS 2013 and 2015 installed. When I tried to compile glew
on a machine which only has VS 2015 installed, it failed because the solution
is for VS 2013. This is fixed by setting the platform toolset to VS 2015 which
allows the VS 2013 solution to be compiled by VS 2015.